Seperate the map view delegate from the controller.
